# commons/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "IOTGlue NodeAPI's",
    "description": "",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/iotglueapi/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/common/rightpanel/sources/{sourceId}": {
      "get": {
        "description": "getSourceForRightPanelById",
        "summary": "getSourceForRightPanelById",
        "tags": [
          "Common - Right Panel"
        ],
        "operationId": "getSourceForRightPanelById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [{
            "name": "sourceId",
            "in": "path",
            "description": "ID of Source",
            "required": true,
            "type": "string"
            }],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },   
...
    "/common/sources/filter": {
      "get": {
        "summary": "getFilteredSources",
        "tags": [
          "Common"
        ],
        "operationId": "getFilteredSources",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mappingReq",
            "in": "query",
            "required": true,
            "type": "boolean",
            "description": ""
          },
          {
            "name": "protocolType",
            "in": "query",
            "required": false,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
},
...
}